Clean up some of the UI tests code
