--- src/Compatibility/ControlGallery/test/Shared.Appium.UITests/TestAttributes.cs
@@ -1,71 +1,0 @@
-﻿namespace UITests
-{
-	public static class Test
-	{
-		public enum Button
-		{
-			Clicked,
-			Command,
-			Text,
-			TextColor,
-			Font,
-			BorderWidth,
-			BorderColor,
-			BorderRadius,
-			Image,
-			Padding,
-			Pressed,
-			LineBreakMode
-		}
-
-		public enum ImageButton
-		{
-			Source,
-			Aspect,
-			IsOpaque,
-			IsLoading,
-			AspectFill,
-			AspectFit,
-			Fill,
-			BorderColor,
-			CornerRadius,
-			BorderWidth,
-			Clicked,
-			Command,
-			Image,
-			Pressed,
-			Padding
-		}
-
-		public enum VisualElement
-		{
-			IsEnabled,
-			Navigation,
-			InputTransparent,
-			Layout,
-			X,
-			Y,
-			AnchorX,
-			AnchorY,
-			TranslationX,
-			TranslationY,
-			Width,
-			Height,
-			Bounds,
-			Rotation,
-			RotationX,
-			RotationY,
-			Scale,
-			IsVisible,
-			Opacity,
-			BackgroundColor,
-			Background,
-			IsFocused,
-			Focus,
-			Unfocus,
-			Focused,
-			Unfocused,
-			Default
-		}
-	}
-}
