--- src/Controls/tests/TestCases.Shared.Tests/Controls - Backup.TestCases.Shared.Tests.csproj
@@ -1,42 +1,0 @@
-﻿<Project Sdk="Microsoft.Build.NoTargets">
-
-  <PropertyGroup>
-    <TargetFrameworks>$(_MauiDotNetTfm)</TargetFrameworks>
-    <GenerateAssemblyInfo>False</GenerateAssemblyInfo>
-    <GenerateMSBuildEditorConfigFile>False</GenerateMSBuildEditorConfigFile>
-    <RootNamespace>Microsoft.Maui.TestCases.Tests</RootNamespace>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <DefineConstants Condition=" '$(ExtraDefineConstants)' != '' ">$(DefineConstants);$(ExtraDefineConstants)</DefineConstants>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Content Include="**\*.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Controls\tests\CustomAttributes\Controls.CustomAttributes.csproj" />
-  </ItemGroup>
-
-  <PropertyGroup Condition=" '$(_MauiTargetPlatformIsAndroid)' == 'True' ">
-    <DefineConstants>$(DefineConstants);ANDROID</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(_MauiTargetPlatformIsiOS)' == 'True' ">
-    <DefineConstants>$(DefineConstants);IOS</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(_MauiTargetPlatformIsMacCatalyst)' == 'True' ">
-    <DefineConstants>$(DefineConstants);MACCATALYST</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(_MauiTargetPlatformIsTizen)' == 'True' ">
-    <DefineConstants>$(DefineConstants);TIZEN</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(_MauiTargetPlatformIsWindows)' == 'True' ">
-    <DefineConstants>$(DefineConstants); WINDOWS</DefineConstants>
-  </PropertyGroup>
-
-</Project>
