+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Concepts/ImageLoadingGalleryTests.cs
@@ -1,5 +1,4 @@
-﻿using Maui.Controls.Sample;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using UITest.Appium;
 using UITest.Core;
 
@@ -35,7 +34,6 @@ namespace Microsoft.Maui.TestCases.Tests
 			remote.GoTo(test.ToString());
 
 			App.WaitForElement($"{test}VisualElement");
-			Thread.Sleep(1000); // android has some button animations that need to finish
 
 			VerifyScreenshot();
 		}
