+++ src/Controls/tests/TestCases.Shared.Tests/Tests/GestureRecognizerUITests.cs
@@ -1,5 +1,4 @@
-﻿using Maui.Controls.Sample;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using NUnit.Framework.Legacy;
 using UITest.Appium;
 using UITest.Core;
