+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Issue21314.cs
@@ -1,4 +1,3 @@
-using Maui.Controls.Sample;
 using NUnit.Framework;
 using NUnit.Framework.Legacy;
 using UITest.Appium;
