+++ src/Controls/tests/TestCases.iOS.Tests/Controls.TestCases.iOS.Tests.csproj
@@ -49,6 +49,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\Controls\tests\CustomAttributes\Controls.CustomAttributes.csproj" />
+    <Using Include="Microsoft.Maui.Controls.CustomAttributes" />
   </ItemGroup>
 
 </Project>
